+++ src/main/java/com/meta/ponkids/domain/system/bbs/repository/BbsRepository.java
@@ -7,15 +7,15 @@ import org.springframework.data.repository.query.Param;
 
 import com.meta.ponkids.domain.system.bbs.entity.Bbs;
 import com.meta.ponkids.domain.system.bbs.repository.custom.BbsRepositoryCustom;
+import com.meta.ponkids.domain.system.ntt.entity.Ntt;
 
 
 
-public interface BbsRepository extends JpaRepository< Bbs, Integer> , BbsRepositoryCustom {
+public interface BbsRepository extends JpaRepository< Bbs, Long> , BbsRepositoryCustom {
 
-   boolean existsByBbsSn(int bbsSn);
+   boolean existsByBbsSn(Long bbsSn);
    
-   public Bbs findByBbsSn(int bbsSn);
-   
+   public Bbs findByBbsSn(Long bbsSn);
    
    @Modifying( clearAutomatically = true )
    @Query( value = "UPDATE tb_bbs "
@@ -23,7 +23,7 @@ public interface BbsRepository extends JpaRepository< Bbs, Integer> , BbsReposit
            + "        , updt_dt = now() "
            + "    WHERE bbs_sn = :bbsSn", nativeQuery = true )
        // nativeQuery true 없으면 error
-   int deleteAllByBbsSn( @Param( "bbsSn" ) int bbsSn );
+   int deleteAllByBbsSn( @Param( "bbsSn" ) Long bbsSn );
 
 	    
 }
